Add new features to allow drawing GeoJSON features on the map
At the core of this, there is a simple framework for drawing
different shapes. The design of it is pretty much the same as in
some of the classes in Rails/ActiveStorage. Currently only
LineStrings are supported, but adding support for other geometry
types is fairly easy.

GpxFile is a helper class for the CLI. It reads a gpx file given
as an argument, and translates it into GeoJSON format. It's only
included in the CLI by default, because when using the Ruby API,
it's more likely that the data is already in GeoJSON format.

Also the version number is now bumped up to 0.1, because this
commit creates new features.